rename file - transforms are stored separately
